add: 1. add FriendSelectViewController for creating a group 2. add groupCreateViewController for creating a group 3. add PickerEditorViewController for create a new picker 4. add PickerSelectionViewController for selecting a picker 5. add PickResultViewController for seeing voting result in time 6. add SearchIDViewController for search valid user and send invitation 7. struct for Group, GroupInfo, Result, User
test:
1. loginViewController for testing

fix:
1. fix some struct

+++ SlobPicker/Helper/DateHelper.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 extension Date {
-    var millisecondsSince1970: Int64 {
-        return Int64((self.timeIntervalSince1970 * 1000.0).rounded())
+    var millisecondsSince1970: Int {
+        return Int((self.timeIntervalSince1970 * 1000.0).rounded())
     }
     
     init(milliseconds: Int64) {
@@ -17,12 +17,8 @@ extension Date {
     }
     
     static var dateFormatter: DateFormatter {
-        
         let formatter = DateFormatter()
-        
         formatter.dateFormat = "yyyy.MM.dd HH:mm"
-                
         return formatter
-        
     }
 }
